Don't crash when collecting runtime metrics on platforms that don't expose ctx switches
`num_ctx_switches` in psutil may raise `NotImplementedError` when it can't find ctx
switches information in the process status file `/proc/$pid/status`:

```
    def num_ctx_switches(self,
                         _ctxsw_re=re.compile(br'ctxt_switches:\t(\d+)')):
        data = self._read_status_file()
        ctxsw = _ctxsw_re.findall(data)
        if not ctxsw:
            raise NotImplementedError(
                "'voluntary_ctxt_switches' and 'nonvoluntary_ctxt_switches'"
                "lines were not found in %s/%s/status; the kernel is "
                "probably older than 2.6.23" % (
                    self._procfs_path, self.pid))
        else:
            return _common.pctxsw(int(ctxsw[0]), int(ctxsw[1]))
```

The comment in the library says that this may happen on kernels older than 2.6.23, but
it also happens when the process is running in Google Cloud Run.

When this happens, it should be safe to ignore and simply not collect any ctx switches
metrics.